# internal/api/api.swagger.yaml
openapi: 3.0.3
info:
  title: Online Courses Platform API
  description: API для платформы онлайн-курсов (аналог Docusaurus + LMS функционал)
  version: 0.1.0
  contact:
    name: Artem
    email: artem@example.com
servers:
  - url: https://api.courses-platform.com/v1
    description: Production server
  - url: http://localhost:3001
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        status:
          type: integer
          nullable: false
        data:
          type: object
          nullable: true
        success:
          type: boolean
          nullable: false
        requestID:
          type: string
          nullable: false
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        fullName:
          type: string
        avatarUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [student, instructor, admin]
        createdAt:
          type: string
          format: date-time
    UserCreate:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
    TokenRefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          description: Refresh токен, полученный при логине
    TokenResponse:
      type: object
      required: [accessToken, refreshToken]
      properties:
        accessToken:
          type: string
          description: Новый access-токен (JWT)
        refreshToken:
          type: string
          description: Новый refresh-токен (opaque token, rotation)
        expiresIn:
          type: integer
          description: Время жизни access-токена в секундах (например 900 = 15 мин)
    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
          minLength: 2
          maxLength: 100
        avatarUrl:
          type: string
          format: uri
          nullable: true
        password:
          type: string
          minLength: 8
          description: Новый пароль (если меняется)
        currentPassword:
          type: string
          minLength: 8
          description: Текущий пароль (обязателен при смене пароля)
    UserDeleteRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          description: Текущий пароль для подтверждения удаления
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        status:
          type: integer
paths:
  /auth/register:
    post:
      operationId: authRegisterUser
      summary: Регистрация нового пользователя
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Некорректные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: authLoginUser
      summary: Авторизация пользователя
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: Успешный вход
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
  /auth/refresh:
    post:
      operationId: authRefreshToken
      summary: Обновление access-токена с помощью refresh-токена (token rotation)
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRefreshRequest"
      responses:
        "200":
          description: Токены успешно обновлены (refresh rotation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Недействительный или истёкший refresh-токен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Некорректный запрос (отсутствует refreshToken)
  /me:
    get:
      operationId: getCurrentUser
      summary: Получить информацию о текущем пользователе
      tags: [Users, Me]
      responses:
        "200":
          description: Информация о пользователе
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Не авторизован

    patch:
      operationId: updateCurrentUser
      summary: Обновить данные текущего пользователя (fullName, avatar, пароль)
      tags: [Users, Me]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Пользователь успешно обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Некорректные данные (например, слабый пароль)
        "401":
          description: Не авторизован
        "403":
          description: Неверный текущий пароль при смене пароля

    delete:
      operationId: deleteCurrentUser
      summary: Удалить аккаунт текущего пользователя (с подтверждением паролем)
      tags: [Users, Me]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDeleteRequest"
      responses:
        "204":
          description: Аккаунт успешно удалён (все данные, курсы, прогресс удаляются)
        "400":
          description: Неверный пароль
        "401":
          description: Не авторизован
...
  /me/courses/{courseID}/lessons/{lessonID}/progress:
    patch:
      operationId: updateLessonProgress
      summary: Обновить прогресс урока
      tags: [Progress]

      parameters:
        - name: courseID
          in: path
          required: true
          schema:
            type: string
        - name: lessonID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                percent:
                  type: number
                completed:
                  type: boolean
                lastWatchedSec:
                  type: integer
      responses:
        "200":
          description: Прогресс обновлен
  /users/{userId}:
    delete:
      operationId: deleteUserById
      summary: Удалить пользователя (только для админов)
      tags: [Users, Admin]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Пользователь удалён
        "403":
          description: Недостаточно прав (только admin)
        "404":
          description: Пользователь не найден
...